working on fixing the tests

+++ Game/game_achievement.cpp
@@ -18,7 +18,7 @@ int32_t game_goal::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_goal::get_error");
-    return (static_cast<int32_t>(game_goal::_last_error));
+    return (game_goal::_last_error);
 }
 
 const char *game_goal::get_error_str() const noexcept
@@ -40,7 +40,7 @@ int32_t game_achievement::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_achievement::get_error");
-    return (static_cast<int32_t>(game_achievement::_last_error));
+    return (game_achievement::_last_error);
 }
 
 const char *game_achievement::get_error_str() const noexcept
