+++ Game/game_behavior_profile.cpp
@@ -462,7 +462,7 @@ int32_t game_behavior_profile::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_behavior_profile::get_error");
-    return (static_cast<int32_t>(game_behavior_profile::_last_error));
+    return (game_behavior_profile::_last_error);
 }
 
 const char *game_behavior_profile::get_error_str() const noexcept
