+++ Game/game_crafting.cpp
@@ -19,7 +19,7 @@ int32_t game_crafting_ingredient::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_crafting_ingredient::get_error");
-    return (static_cast<int32_t>(game_crafting_ingredient::_last_error));
+    return (game_crafting_ingredient::_last_error);
 }
 
 const char *game_crafting_ingredient::get_error_str() const noexcept
@@ -41,7 +41,7 @@ int32_t game_crafting::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_crafting::get_error");
-    return (static_cast<int32_t>(game_crafting::_last_error));
+    return (game_crafting::_last_error);
 }
 
 const char *game_crafting::get_error_str() const noexcept
