+++ Game/game_currency_rate.cpp
@@ -368,7 +368,7 @@ int32_t game_currency_rate::get_error() const noexcept
 {
     errno_abort_if_uninitialised(this->_initialised_state,
         "game_currency_rate::get_error");
-    return (static_cast<int32_t>(game_currency_rate::_last_error));
+    return (game_currency_rate::_last_error);
 }
 
 const char *game_currency_rate::get_error_str() const noexcept
