+++ Game/game_dialogue_line.cpp
@@ -505,7 +505,7 @@ int32_t game_dialogue_line::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_dialogue_line::get_error");
-    return (static_cast<int32_t>(game_dialogue_line::_last_error));
+    return (game_dialogue_line::_last_error);
 }
 
 const char *game_dialogue_line::get_error_str() const noexcept
