+++ Game/game_economy_table.cpp
@@ -674,7 +674,7 @@ int32_t game_economy_table::get_error() const noexcept
 {
     errno_abort_if_uninitialised(this->_initialised_state,
         "game_economy_table::get_error");
-    return (static_cast<int32_t>(game_economy_table::_last_error));
+    return (game_economy_table::_last_error);
 }
 
 const char *game_economy_table::get_error_str() const noexcept
