+++ Game/game_equipment.cpp
@@ -415,7 +415,7 @@ int32_t game_equipment::get_error() const noexcept
 {
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state, "game_equipment::get_error");
-    return (static_cast<int32_t>(game_equipment::_last_error));
+    return (game_equipment::_last_error);
 }
 
 const char *game_equipment::get_error_str() const noexcept
