+++ Game/game_event_scheduler.cpp
@@ -616,7 +616,7 @@ int32_t game_event_scheduler::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_event_scheduler::get_error");
-    return (static_cast<int32_t>(game_event_scheduler::_last_error));
+    return (game_event_scheduler::_last_error);
 }
 
 const char *game_event_scheduler::get_error_str() const noexcept
