+++ Game/game_map3d.cpp
@@ -96,7 +96,7 @@ int32_t game_map3d::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_map3d::get_error");
-    return (static_cast<int32_t>(game_map3d::_last_error));
+    return (game_map3d::_last_error);
 }
 
 const char *game_map3d::get_error_str() const noexcept
