+++ Game/game_progress_tracker.cpp
@@ -18,7 +18,7 @@ int32_t game_progress_tracker::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_progress_tracker::get_error");
-    return (static_cast<int32_t>(game_progress_tracker::_last_error));
+    return (game_progress_tracker::_last_error);
 }
 
 const char *game_progress_tracker::get_error_str() const noexcept
