+++ Game/game_quest.cpp
@@ -17,7 +17,7 @@ int32_t    game_quest::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_quest::get_error");
-    return (static_cast<int32_t>(game_quest::_last_error));
+    return (game_quest::_last_error);
 }
 
 const char    *game_quest::get_error_str() const noexcept
