+++ Game/game_rarity_band.cpp
@@ -348,7 +348,7 @@ int32_t game_rarity_band::get_error() const noexcept
 {
     errno_abort_if_uninitialised(this->_initialised_state,
         "game_rarity_band::get_error");
-    return (static_cast<int32_t>(game_rarity_band::_last_error));
+    return (game_rarity_band::_last_error);
 }
 
 const char *game_rarity_band::get_error_str() const noexcept
