+++ Game/game_resistance.cpp
@@ -365,7 +365,7 @@ int32_t game_resistance::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_resistance::get_error");
-    return (static_cast<int32_t>(game_resistance::_last_error));
+    return (game_resistance::_last_error);
 }
 
 const char *game_resistance::get_error_str() const noexcept
