+++ Game/game_scripting_bridge.cpp
@@ -339,7 +339,7 @@ int32_t game_script_context::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_script_context::get_error");
-    return (static_cast<int32_t>(game_script_context::_last_error));
+    return (game_script_context::_last_error);
 }
 
 const char *game_script_context::get_error_str() const noexcept
@@ -1400,7 +1400,7 @@ int32_t game_script_bridge::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_script_bridge::get_error");
-    return (static_cast<int32_t>(game_script_bridge::_last_error));
+    return (game_script_bridge::_last_error);
 }
 
 const char *game_script_bridge::get_error_str() const noexcept
