+++ Game/game_server.cpp
@@ -19,7 +19,7 @@ int32_t game_server::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_server::get_error");
-    return (static_cast<int32_t>(game_server::_last_error));
+    return (game_server::_last_error);
 }
 
 const char *game_server::get_error_str() const noexcept
