+++ Game/game_state.cpp
@@ -617,7 +617,7 @@ int32_t game_state::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_state::get_error");
-    return (static_cast<int32_t>(game_state::_last_error));
+    return (game_state::_last_error);
 }
 
 const char *game_state::get_error_str() const noexcept
