+++ Game/game_vendor_profile.cpp
@@ -498,7 +498,7 @@ int32_t game_vendor_profile::get_error() const noexcept
 {
     errno_abort_if_uninitialised(this->_initialised_state,
         "game_vendor_profile::get_error");
-    return (static_cast<int32_t>(game_vendor_profile::_last_error));
+    return (game_vendor_profile::_last_error);
 }
 
 const char *game_vendor_profile::get_error_str() const noexcept
