+++ Game/game_world.cpp
@@ -889,7 +889,7 @@ int32_t game_world::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised_or_destroyed(this->_initialised_state,
             "game_world::get_error");
-    return (static_cast<int32_t>(game_world::_last_error));
+    return (game_world::_last_error);
 }
 
 const char *game_world::get_error_str() const noexcept
